Removed the default identifier value constraint from Entity
The container in the Entity classes constructor that constrained the Entity identifier to be non-default has been removed. This is done to make the Entity class more flexible.
When creating a new entity, where the data storage solution is responsible for creating the unique identifier, it's impractical to not be allowed to create an entity with a default value identifier. Whatever homemade default value that's attached to the new entity are not going to be valid anyways.