Reimplemented Value using reflection.
Usage of the Value class has been greatly simplified because it now uses reflection to identifier fields and properties used in equality checks instead of relying on a list of properties manually provided by the user of the library. By default all properties and field are included in equality checks. Field and properties can be annotated with attributes to exclude (and include) them in equality checks.

Using reflection instead of a manually implemented list of fields and properties also greatly reduces the chance of introducing bugs, if a field or property is added to a value, but not the list if fields and properties used in equality checks.

See the readme for more details.